requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
pyyaml
Pillow
scikit-learn
seaborn